// src/payload-types.ts
export interface Pool {
  id: string;
  chainId: number;
  network: string;
  address?: string | null;
  name: string;
  team: string;
  treasury: string;
  paymentRules: {
    index: number;
    paymentToken: {
      address: string;
      name?: string | null;
      symbol: string;
      decimals: number;
    };
    price: string;
    minPurchase: string;
    maxPurchase: string;
    enabled: boolean;
    id?: string | null;
  }[];
  saleToken: {
    address: string;
    name?: string | null;
    symbol: string;
    decimals: number;
  };
  status:
    | "ACTIVE"
    | "PAUSED"
    | "STOPPED"
    | "SOLD_OUT"
    | "CREATING"
    | "UPCOMING";
  isHidden: boolean;
  apr: number;
  deployedAt: number;
  saleStartedAt: number;
  saleDuration: number;
  lockDuration: number;
  pausedDurationSum: number;
  totalSaleCap: string;
  totalSold: string;
  creator: string;
  createdHash?: string | null;
  updatedAt: string;
  createdAt: string;
}